+++ lib/testcentricity_web/web_elements/radio.rb
@@ -2,11 +2,13 @@ module TestCentricity
   class Radio < UIElement
     attr_accessor :proxy
     attr_accessor :label
+    attr_accessor :input
 
     def initialize(name, parent, locator, context)
       super
       @type = :radio
       radio_spec = {
+        input: nil,
         proxy: nil,
         label: nil
       }
@@ -16,6 +18,8 @@ module TestCentricity
     def define_custom_elements(element_spec)
       element_spec.each do |element, value|
         case element
+        when :input
+          @input = value
         when :proxy
           @proxy = value
         when :label
@@ -44,9 +48,14 @@ module TestCentricity
     #   accept_terms_radio.selected?
     #
     def selected?
-      obj, = find_element(:all)
-      object_not_found_exception(obj, 'Radio')
-      obj.checked?
+      @base_object, = find_element(:all)
+      object_not_found_exception(@base_object, 'Radio')
+      rad = if @input.nil?
+              @base_object
+            else
+              find_component(@input, 'Radio')
+            end
+      rad.checked?
     end
 
     # Is radio button visible?
@@ -56,7 +65,14 @@ module TestCentricity
     #   accept_terms_radio.visible?
     #
     def visible?
-      @proxy.nil? ? super : page.find(:css, @proxy).visible?
+      if @proxy.nil?
+        super
+      else
+        @base_object, = find_element(:all)
+        object_not_found_exception(@base_object, 'Radio')
+        proxy = find_component(@proxy, 'radio proxy')
+        proxy.visible?
+      end
     end
 
     # Is radio button disabled (not enabled)?
@@ -66,10 +82,14 @@ module TestCentricity
     #   accept_terms_radio.disabled?
     #
     def disabled?
-      visibility = @proxy.nil? ? true : :all
-      obj, type = find_element(visibility)
-      object_not_found_exception(obj, type)
-      obj.disabled?
+      if @input.nil?
+        super
+      else
+        @base_object, = find_element(:all)
+        object_not_found_exception(@base_object, 'Radio')
+        rad = find_component(@input, 'radio')
+        rad.disabled?
+      end
     end
 
     # Return radio button caption
@@ -80,9 +100,15 @@ module TestCentricity
     #
     def get_value
       if @label.nil?
-        @proxy.nil? ? super : page.find(:css, @proxy).text
+        if @proxy.nil?
+          super
+        else
+          proxy = find_component(@proxy, 'radio proxy')
+          proxy.text
+        end
       else
-        page.find(:css, @label).text
+        label = find_component(@label, 'radio label')
+        label.text
       end
     end
 
@@ -97,17 +123,22 @@ module TestCentricity
     #   accept_terms_radio.set_selected_state(true)
     #
     def set_selected_state(state)
-      obj, = find_element(:all)
-      object_not_found_exception(obj, 'Radio')
-      invalid_object_type_exception(obj, 'radio')
-      if @proxy.nil?
-        begin
-          obj.set(state)
-        rescue
-          obj.click unless state == obj.checked?
+      @base_object, = find_element(:all)
+      object_not_found_exception(@base_object, 'Radio')
+      proxy = find_component(@proxy, 'radio proxy') unless @proxy.nil?
+      rad = find_component(@input, 'radio') unless @input.nil?
+      if @input.nil?
+        if @proxy.nil?
+          @base_object.click unless state == @base_object.checked?
+        else
+          proxy.click unless state == @base_object.checked?
         end
       else
-        page.find(:css, @proxy).click unless state == obj.checked?
+        if @proxy.nil?
+          @base_object.click unless state == rad.checked?
+        else
+          proxy.click unless state == rad.checked?
+        end
       end
     end